# back/api/openapi.yaml
openapi: 3.0.3
info:
  title: upcontrol API
  version: 1.0.0
  description: |
    The front-facing contract for upcontrol. Generated into a strict Go server
    (oapi-codegen) and into TS types for `front/src/lib/api.d.ts`. Response
    shapes are pinned to `front/src/lib/mockData.ts` — a contract test
    (build-plan §2.7) checks each against golden JSON taken from the mocks.

    Rules from the plan (§4.4):
      - Version in the path (/v1), not a header.
      - Errors are one shape: `{error: {code, message, upgrade?}}`. The optional
      - Errors are one shape: `{error: {code, message, upgrade?}}`. The optional
        text reason. No `disabled`, no `navigate('/plan')`.
      - Pagination only on unbounded lists: incidents, logs. Monitors, channels
        and recipients are plan-bounded and returned whole.
      - `null` is never used for "no value": the field is omitted. "Zero is
        silence" — an absent field does not render.
      - All timestamps are RFC 3339 UTC.
      - `POST /i` and `POST /hooks/{provider}` are `x-hand-written: true`: they
        exist here for the /docs/api page but are implemented by hand (the
        generator's validation is forbidden on /i by §4.5).
servers:
  - url: http://localhost:5173
    description: local dev (front dev server proxies to ucapi)

# ---------------------------------------------------------------------------
# Common building blocks
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: uc_session
    apiKey:
      type: apiKey
      in: header
      name: X-Upcontrol-Key
  responses:
...
    Unauthorized:
      description: No session, or the key is not ours.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: "plan_limit_exceeded" }
            message: { type: string }
            upgrade:
              type: object
              description: Present only on 402. Drives the client's upgrade prompt.
              required: [reason]
              properties:
                reason:
                  type: string
                  example: "Free allows 3 HTTP checks."
                highlight:
                  type: string
                  description: Optional row to highlight on the pricing table.

    # --- account ------------------------------------------------------------
...
    ApiKey:
      type: object
      required: [id, prefix, createdAt]
      description: |
        The active API key. The full secret is never stored and never returned
        by GET /v1/keys — only this identifier (the prefix). The full key is
        returned exactly once by POST /v1/keys/rotate.
      properties:
        id: { type: string, example: "key_1" }
        prefix: { type: string, example: "uc_live_8f2ac41d9b0e" }
        createdAt: { type: string, format: date-time }
    RotatedKey:
      type: object
      required: [id, prefix, createdAt, value]
      description: |
        POST /v1/keys/rotate's answer — the ONE place the full key ever
        appears. `value` is shown exactly once and is not retrievable again;
        `prefix` is what GET /v1/keys lists from here on.
      properties:
        id: { type: string, example: "key_2" }
        prefix: { type: string, example: "uc_live_8f2ac41d9b0e" }
        createdAt: { type: string, format: date-time }
        value: { type: string, example: "uc_live_8f2ac41d9b0e6c31a57f92d4" }
    KeyUsageEntry:
      type: object
      required: [time, endpoint, status]
      properties:
        time: { type: string, example: "15:12" }
        endpoint: { type: string, example: "POST /v1/event" }
        status: { type: integer, example: 202 }
    KeysResponse:
      type: object
      required: [key, usage]
      properties:
        key: { $ref: "#/components/schemas/ApiKey" }
        usage:
          type: array
          items: { $ref: "#/components/schemas/KeyUsageEntry" }

    # --- channels / recipients ---------------------------------------------
...
paths:
...
  /v1/keys:
    get:
      tags: [sources]
      summary: The active API key + recent usage (Sources #key).
      responses:
        "200":
          description: OK
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/KeysResponse" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/keys/rotate:
    post:
      tags: [sources]
      summary: |
        Rotate the active key. The old key stays valid for a 24h overlap window,
        then 401s — a rotation that kills a deployed app fails at the customer.
      responses:
        "200":
          description: Rotated; the new key is returned.
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/RotatedKey" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/instance/ai:
    put:
      tags: [instance]
      summary: >-
        Self-host only (the hosted cloud answers 404): store the instance's
        AI settings — API key, model, base URL — each optional, so changing
        the model never demands re-pasting the key. Values are sealed
        (AES-256-GCM under UC_SECRET_KEY_HEX) before they land in Postgres
        and take effect on the next Explain — no restart. Nothing here is
        ever returned by any endpoint; presence is read through the explain
        preview's `model`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                  description: >-
                    An OpenAI-format API key (like sk-...). Any
                    OpenAI-compatible provider's key works.
                model:
                  type: string
                  description: The model name as the provider spells it, e.g. gpt-4o-mini.
                baseUrl:
                  type: string
                  description: >-
                    The chat-completions endpoint's base, e.g.
                    https://api.openai.com/v1 — any OpenAI-compatible proxy
                    or gateway. Empty keeps the env-configured one.
      responses:
        "204": { description: Stored. }
        "400": { description: invalid_key / invalid_model / invalid_base_url / nothing_to_store — the message names the fix. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: notify_role — changing the instance's brain is a login-role act. }
        "404": { description: Not a self-host — this door does not exist on the hosted cloud. }
        "503": { description: secret_key_missing — no UC_SECRET_KEY_HEX to seal with. }
    delete:
      tags: [instance]
      summary: >-
        Self-host only: remove every Settings-set AI value. Env config, if
        any, takes over; otherwise Explain turns off.
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: notify_role }
        "404": { description: Not a self-host. }
  /v1/instance/telegram-bot:
    put:
      tags: [instance]
      summary: >-
        Self-host only (the hosted cloud answers 404): store the Telegram
        bot's token and username. Sealed like the AI key. Alerts pick the
        token up on the next send and the Channels/recipients surfaces
        appear immediately; the bot's /start linking and Mini App sign-in
        bind at boot, so they follow on the next restart.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, username]
              properties:
                token:
                  type: string
                  description: The token exactly as @BotFather printed it.
                username:
                  type: string
                  description: The bot's username, with or without the leading @.
      responses:
        "204": { description: Stored. }
        "400": { description: invalid_token / invalid_username — shape refusals with the fix in the message. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: notify_role }
        "404": { description: Not a self-host. }
        "503": { description: secret_key_missing — no UC_SECRET_KEY_HEX to seal with. }
    delete:
      tags: [instance]
      summary: >-
        Self-host only: remove the Settings-set bot token and username. Env
        config, if any, takes over; otherwise the Telegram surface folds
        away.
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: notify_role }
        "404": { description: Not a self-host. }
  /v1/instance/smtp:
    put:
      tags: [instance]
      summary: >-
        Self-host only (the hosted cloud answers 404): store the SMTP relay
        for sign-in mail and email alerts. Each field optional — send only
        what changes — sealed like the AI key, resolved per message, so a
        saved relay works with no restart. Values are never returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                host:
                  type: string
                  description: Relay hostname alone (no scheme, no port), like smtp.eu.mailgun.org.
                port:
                  type: string
                  description: "Relay port as a string; usually 587 (STARTTLS) or 465. Default: 587."
                username:
                  type: string
                  description: Relay login, often the sending address itself.
                password:
                  type: string
                  description: Relay password or app password.
                from:
                  type: string
                  description: The From address; most relays insist it is a verified one.
      responses:
        "204": { description: Stored. }
        "400": { description: invalid_host / invalid_port / invalid_username / invalid_password / invalid_from / nothing_to_store — refusals with the fix in the message. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: notify_role }
        "404": { description: Not a self-host. }
        "503": { description: secret_key_missing — no UC_SECRET_KEY_HEX to seal with. }
    delete:
      tags: [instance]
      summary: >-
        Self-host only: remove every Settings-set SMTP value. Env config
        (UC_SMTP_*), if any, takes over; otherwise email is off and sign-in
        codes land in the ucapi log again.
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: notify_role }
        "404": { description: Not a self-host. }
...