# back/gen/pg/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package pg

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type AiCall struct {
	ID               int64
	TenantID         int64
	Scenario         string
	Model            string
	PromptTokens     int64
	CompletionTokens int64
	CreatedAt        pgtype.Timestamptz
}

type AiExplainCache struct {
	TenantID  int64
	InputHash []byte
	Text      string
	CreatedAt pgtype.Timestamptz
}

type AiUsage struct {
	TenantID         int64
	Month            pgtype.Date
	Used             int32
	PromptTokens     int64
	CompletionTokens int64
}

type AlertChannel struct {
	ID                int64
	PublicID          pgtype.UUID
	TenantID          int64
	Kind              string
	Target            string
	SecretEnc         []byte
	BreakerOpenUntil  pgtype.Timestamptz
	CreatedAt         pgtype.Timestamptz
	Notify            []byte
	RecipientPersonID *int64
	MutedUntil        pgtype.Timestamptz
}

type ApiKey struct {
	ID            int64
	TenantID      int64
	ProjectID     int64
	Prefix        string
	SecretHash    []byte
	State         string
	RotatingUntil pgtype.Timestamptz
	CreatedAt     pgtype.Timestamptz
	LastUsedAt    pgtype.Timestamptz
}

type BillingSubscription struct {
	TenantID         int64
	LsCustomerID     int64
	LsSubscriptionID int64
	VariantID        int64
	Status           string
	RenewsAt         pgtype.Timestamptz
	EndsAt           pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
	CreatedAt        pgtype.Timestamptz
}

type DeliveryAttempt struct {
	ID      int64
	QueueID int64
	At      pgtype.Timestamptz
	Outcome string
	Detail  string
}

type DeliveryQueue struct {
	ID         int64
	TenantID   int64
	IncidentID *int64
	ChannelID  int64
	IdemKey    string
	Class      string
	Payload    []byte
	Attempts   int32
	NextTryAt  pgtype.Timestamptz
	State      string
	DeadReason *string
	LeasedBy   *string
	LeaseUntil pgtype.Timestamptz
}

type ErrorAlertState struct {
	TenantID    int64
	Fingerprint int64
	Kind        string
	LastAlerted pgtype.Timestamptz
}

type Incident struct {
	ID            int64
	PublicID      pgtype.UUID
	TenantID      int64
	ProjectID     int64
	MonitorID     *int64
	Detector      string
	Fingerprint   int64
	Title         string
	Status        string
	DetectedAt    pgtype.Timestamptz
	NotifiedAt    pgtype.Timestamptz
	AckedAt       pgtype.Timestamptz
	AckedBy       *int64
	ResolvedAt    pgtype.Timestamptz
	CloseReason   *string
	AffectedCount int32
	DeployID      *int64
	SlicePhase    int16
	SliceDoneAt   pgtype.Timestamptz
}

type IncidentSlice struct {
	IncidentID int64
	Seq        int64
	Ts         pgtype.Timestamptz
	Level      string
	Service    string
	Message    string
}

type IncidentUpdate struct {
	ID         int64
	IncidentID int64
	At         pgtype.Timestamptz
	Kind       string
	Text       string
}

type IngestBatch struct {
	BatchKey   string
	BodyHash   []byte
	AcceptedAt pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
	Accepted   int32
}

type InstallToken struct {
	ID        int64
	TenantID  int64
	ProjectID int64
	TokenHash []byte
	CreatedAt pgtype.Timestamptz
	ExpiresAt pgtype.Timestamptz
	UsedAt    pgtype.Timestamptz
}

type InstanceSetting struct {
	Key       string
	ValueEnc  []byte
	UpdatedAt pgtype.Timestamptz
}

type KeyUsageLog struct {
	ID       int64
	TenantID int64
	KeyID    int64
	At       pgtype.Timestamptz
	Source   string
	Outcome  string
}

type MagicLinkCode struct {
	Email      string
	CodeHash   []byte
	Attempts   int32
	ExpiresAt  pgtype.Timestamptz
	RedeemedAt pgtype.Timestamptz
	CreatedAt  pgtype.Timestamptz
}

type MagicLinkIp struct {
	Ip      string
	FirstAt pgtype.Timestamptz
	Count   int32
}

type Monitor struct {
	ID                 int64
	PublicID           pgtype.UUID
	TenantID           int64
	ProjectID          int64
	Kind               string
	Name               string
	Target             string
	Keyword            *string
	IntervalSec        int32
	AvailabilityTarget pgtype.Numeric
	Paused             bool
	PingToken          *string
	GraceSec           *int32
	CreatedAt          pgtype.Timestamptz
}

type MonitorFact struct {
	MonitorID           int64
	Status              string
	SslExpiresAt        pgtype.Timestamptz
	DomainExpiresAt     pgtype.Timestamptz
	LastCheckAt         pgtype.Timestamptz
	ConsecutiveFailures int32
}

type MonitorSchedule struct {
	MonitorID  int64
	Region     string
	NextDueAt  pgtype.Timestamptz
	LeasedBy   *string
	LeaseUntil pgtype.Timestamptz
}

type Person struct {
	ID              int64
	PublicID        pgtype.UUID
	Email           *string
	EmailVerifiedAt pgtype.Timestamptz
	TelegramID      *int64
	GoogleSub       *string
	Name            string
	CreatedAt       pgtype.Timestamptz
}

type PlanEntitlement struct {
	Plan               string
	HttpChecks         int32
	Regions            int32
	WindowLines        int64
	WindowHours        int32
	RetainMult         pgtype.Numeric
	AiExplains         *int32
	IncidentDays       int32
	MinIntervalSec     int32
	TelegramRecipients int32
}

type ProbeNode struct {
	ID         string
	Region     string
	LastSeenAt pgtype.Timestamptz
	BlindSince pgtype.Timestamptz
}

type Project struct {
	ID        int64
	PublicID  pgtype.UUID
	TenantID  int64
	Domain    string
	CreatedAt pgtype.Timestamptz
	Meta      []byte
}

type ProjectSeq struct {
	ProjectID int64
	Next      int64
}

type ProjectWindow struct {
	ProjectID    int64
	CutoffSeq    int64
	RetainSeq    int64
	WindowHours  pgtype.Numeric
	BeyondErrors *int64
	ComputedAt   pgtype.Timestamptz
}

type Session struct {
	ID         int64
	TokenHash  []byte
	PersonID   int64
	TenantID   int64
	CreatedAt  pgtype.Timestamptz
	LastSeenAt pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
}

type SourceConnection struct {
	ID           int64
	TenantID     int64
	ProjectID    int64
	Kind         string
	ExternalID   *string
	TokenEnc     []byte
	Status       string
	LastSignalAt pgtype.Timestamptz
	Paused       bool
	HookToken    *string
	LastEvent    *string
}

type StatusPage struct {
	ID               int64
	TenantID         int64
	ProjectID        int64
	Slug             string
	Domain           *string
	DomainVerifiedAt pgtype.Timestamptz
	Title            string
	Components       []byte
	Config           []byte
}

type TelegramInvite struct {
	ID         int64
	TenantID   int64
	Role       string
	InvitedBy  int64
	TokenHash  []byte
	CreatedAt  pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
	RedeemedAt pgtype.Timestamptz
}

type Tenant struct {
	ID             int64
	PublicID       pgtype.UUID
	Name           string
	Plan           string
	Billing        string
	CreatedAt      pgtype.Timestamptz
	ClaimTokenHash []byte
	ClaimedAt      pgtype.Timestamptz
}

type TenantLineLedger struct {
	ProjectID int64
	Bucket5m  pgtype.Timestamptz
	Rows      int64
	MinSeq    int64
	MaxSeq    int64
}

type TenantMember struct {
	TenantID int64
	PersonID int64
	Role     string
	Status   string
}

type WebVisitor struct {
	ID               int64
	TokenHash        []byte
	FirstSeenAt      pgtype.Timestamptz
	LastSeenAt       pgtype.Timestamptz
	FirstReferrer    string
	FirstUtmSource   string
	FirstUtmMedium   string
	FirstUtmCampaign string
	FirstCountry     string
	FirstDevice      string
	FirstPath        string
	LastCountry      string
	LastDevice       string
	Email            string
	PersonID         *int64
	TenantID         *int64
	SignedInAt       pgtype.Timestamptz
	AccountCreatedAt pgtype.Timestamptz
	EventsCount      int64
	IsBot            bool
}

type WebhookSeen struct {
	Provider string
	EventID  string
	SeenAt   pgtype.Timestamptz
}